:hammer: BaseESN abc + template method pattern
