+++ esnpy/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from .esn import ReservoirConfig, ESN, DeepESN
+from .esn import ESN, DeepESN
+from .reservoir import ReservoirConfig
 from . import init
 from . import tune
 from . import train
